Flask app/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
psycopg2-binary==2.9.3